Remove C++ templates
