+++ src/Functions/toBool.cpp
@@ -1,6 +1,5 @@
 #include <Functions/IFunction.h>
 #include <Functions/FunctionFactory.h>
-#include <Functions/FunctionsConversion.h>
 #include <Functions/CastOverloadResolver.h>
 #include <DataTypes/DataTypeFactory.h>
 #include <DataTypes/DataTypeString.h>
@@ -12,6 +11,14 @@ namespace
 {
     class FunctionToBool : public IFunction
     {
+    private:
+        ContextPtr context;
+
+        static String getReturnTypeName(const DataTypePtr & argument)
+        {
+            return argument->isNullable() ? "Nullable(Bool)" : "Bool";
+        }
+
     public:
         static constexpr auto name = "toBool";
 
@@ -32,8 +39,7 @@ namespace
 
         DataTypePtr getReturnTypeImpl(const DataTypes & arguments) const override
         {
-            auto bool_type = DataTypeFactory::instance().get("Bool");
-            return arguments[0]->isNullable() ? makeNullable(bool_type) : bool_type;
+            return DataTypeFactory::instance().get(getReturnTypeName(arguments[0]));
         }
 
         ColumnPtr executeImpl(const ColumnsWithTypeAndName & arguments, const DataTypePtr & result_type, size_t) const override
@@ -42,18 +48,17 @@ namespace
             {
                 arguments[0],
                 {
-                    DataTypeString().createColumnConst(arguments[0].column->size(), arguments[0].type->isNullable() ? "Nullable(Bool)" : "Bool"),
+                    DataTypeString().createColumnConst(arguments[0].column->size(), getReturnTypeName(arguments[0].type)),
                     std::make_shared<DataTypeString>(),
                     ""
                 }
             };
 
-            FunctionOverloadResolverPtr func_builder_cast = CastInternalOverloadResolver<CastType::nonAccurate>::createImpl();
+            FunctionOverloadResolverPtr func_builder_cast = createInternalCastOverloadResolver(CastType::nonAccurate, {});
             auto func_cast = func_builder_cast->build(cast_args);
             return func_cast->execute(cast_args, result_type, arguments[0].column->size());
         }
     };
-
 }
 
 REGISTER_FUNCTION(ToBool)
