+++ src/Storages/PartitionedSink.cpp
@@ -2,8 +2,6 @@
 
 #include <Common/ArenaUtils.h>
 
-#include <Functions/FunctionsConversion.h>
-
 #include <Interpreters/Context.h>
 #include <Interpreters/ExpressionAnalyzer.h>
 #include <Interpreters/TreeRewriter.h>
@@ -34,7 +32,7 @@ PartitionedSink::PartitionedSink(
     , sample_block(sample_block_)
 {
     ASTs arguments(1, partition_by);
-    ASTPtr partition_by_string = makeASTFunction(FunctionToString::name, std::move(arguments));
+    ASTPtr partition_by_string = makeASTFunction("toString", std::move(arguments));
 
     auto syntax_result = TreeRewriter(context).analyze(partition_by_string, sample_block.getNamesAndTypesList());
     partition_by_expr = ExpressionAnalyzer(partition_by_string, syntax_result, context).getActions(false);
